Fixes to allow compilation using f95.
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@846 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/image_mod.f90
@@ -625,7 +625,7 @@ thisPVimage%slitDirection - (thisPVimage%slitWidth/2.)*slitnorm
        type(IMAGETYPE) :: image(:)
        integer, allocatable :: rImage(:,:), gImage(:,:), bImage(:,:)
        integer :: nx, ny
-       real :: f5, f95
+       real(double) :: f5, f95
        real :: t
        integer :: i,j, k, nsize
